Lake manifest change
For reasons unknown to me, this leads to better caching behavior.
Without the change, from a fresh docker install, and
```
lake exe cache get
lake build
```
in this directory, I experience the second line rebuilding all of
mathlib. With the change, I see `mathlib4` correctly downloaded
and reused. @dwrensha was not able to repro this behavior on CI,
so I don't know what's going on.