A better definition of affine Rupert sets.
I think it is reasonable to restrict attention to finite-dimensional
spaces. I frankly have little mathematical confidence with
infinite-dimensional linear algebra. Adding this assumption means that
we can synthesize an instance of `Q.direction.HasOrthogonalProjection`
after all, as desired.

More explanation:

Recall that an affine space is a "vector space that has forgotten its
origin", and any given affine space P is defined with respect to a
vector space V that constitutes the directional steps one can take in
P. The "direction" of an affine subspace Q ⊆ P is the subspace W ⊆ V
such that W is all of the steps one can take while staying inside Q.
`Q.direction.HasOrthogonalProjection` asserts that we can in fact
carry out the process of orthogonal projection onto the direction of
the affine subspace Q. In a finite-dimensional space, this is not
hard. Were we working in an infinite-dimensional space, I think we'd
need to require completeness of V and/or P, and also I believe we'd
need to prove that Q.direction inherits that completeness. I didn't
succeed in finding any way to get the typeclass to be instantiated
from what's already in mathlib.